+++ src/network/FixedLengthString.h
@@ -14,12 +14,12 @@ class FixedLengthString
 	 * @param[in] size size of the string
 	 * @param[in] value string of the size size
 	 */ 
-        FixedLengthString(const char* value, uint8_t size);
+        FixedLengthString(const char* value, size_t size);
         /**
          * create an empty string
          * @param[în] size size of the string
          */
-        FixedLengthString(uint8_t size);
+        FixedLengthString(size_t size);
 	/**
 	 * copy constructor
 	 * @param[in] other instance to copy
@@ -34,7 +34,7 @@ class FixedLengthString
 	 * getSize get the size of the string returned by getString()
 	 * @return size of string
 	 */
-	uint8_t getSize();
+	size_t getSize();
 	/**
 	 * copy and convert this object to a standard string
 	 */
@@ -45,7 +45,7 @@ class FixedLengthString
 	 * @return reference to the selected index
 	 * @pre index > 0 && index < this->getLength()
 	 */
-	char& operator[](uint8_t index);
+	char& operator[](size_t index);
         /**
          * Compare two fixed length string for equality
          *
@@ -58,7 +58,7 @@ class FixedLengthString
 	 */
 	virtual ~FixedLengthString();
     private:
-	uint8_t size;
+	size_t size;
 	char* value;
 };
 #endif
